Move the music extension into a new extensions directory
